rename: 정회원 관련 컴포넌트 AllMembers 폴더로 이동
